moves menu item to gene glyph
